Fix: remove old routes
