+++ apps/web-client/src/routes/inventory/+layout.svelte
