+++ apps/web-client/src/routes/inventory/+page.ts
@@ -5,7 +5,7 @@ import { base } from "$app/paths";
 import { appPath } from "$lib/paths";
 
 export const load: Load = async ({ url }) => {
-	if ([`${base}/proto/inventory`, `${base}/proto/inventory/`].includes(url.pathname)) {
+	if ([`${base}/inventory`, `${base}/inventory/`].includes(url.pathname)) {
 		throw redirect(307, appPath("warehouses"));
 	}
 };
