+++ apps/web-client/src/routes/inventory/warehouses/+page.svelte
