+++ apps/web-client/src/routes/inventory/warehouses/[...id]/+page.svelte
