+++ apps/web-client/src/routes/outbound/[...id]/+page.svelte
