+++ apps/web-client/src/routes/stock/+page.svelte
